api/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cloudinary==1.30.0
Django==4.1.5
django-autoslug==1.9.8
django-cors-headers==3.13.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==16.6.1
flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
